--- eventmesh-storage/eventmesh-storage-knative/src/main/java/org/apache/eventmesh/connector/knative/cloudevent/impl/KnativeHeaders.java
@@ -1,31 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.eventmesh.connector.knative.cloudevent.impl;
-
-public class KnativeHeaders {
-
-    public static final String CONTENT_TYPE = "Content-Type";
-
-    public static final String CE_ID = "Ce-Id";
-
-    public static final String CE_SPECVERSION = "Ce-Specversion";
-
-    public static final String CE_TYPE = "Ce-Type";
-
-    public static final String CE_SOURCE = "Ce-Source";
-}
